Update App.js
So far, this contains the login page and its basic components except for a possible button that will lead to an account creation page.
This is the first page that will be displayed when the app is opened. All it does for now, is print "You have entered the correct name and password X times.", every time the enter button is pressed with "hello" entered as the username, and "goodbye" entered as the password, X being a variable that is incremented with every press.